Now using rspec_hpricot_matchers and titles, not ids
Using the hpricot matchers is _way_ easier than regexes.

Using titles not ids seems more natural, and will also mean that
externally, it doesn't matter what my backend is (after all, if/when
a git backend is used, there won't be ids in all likelihood).

This spec re-write is also trying to keep to the philosophy of 'only
external results matter'.  There are no specs of the model, and
validations that are checked, are checked by the error codes or
otherwise returned.